Split message analyser into multiple services
