+++ IPBot/IPBot/Interfaces/IAnimeAnalyserService.cs
@@ -1,0 +1,6 @@
+﻿namespace IPBot.Interfaces;
+
+public interface IAnimeAnalyserService
+{
+    Task CheckMessageForAnimeAsync(SocketMessage message);
+}
