+++ IPBot/IPBot/Interfaces/IHatefulContentAnalyserService.cs
@@ -1,0 +1,6 @@
+﻿namespace IPBot.Interfaces;
+
+public interface IHatefulContentAnalyserService
+{
+    Task CheckMessageForHatefulContentAsync(SocketMessage message);
+}
