+++ IPBot/IPBot/Interfaces/IMessageMediaAnalyserService.cs
@@ -1,0 +1,6 @@
+﻿namespace IPBot.Interfaces;
+
+public interface IMessageMediaAnalyserService
+{
+    Task<List<string>> GetContentUrlsAsync(SocketMessage message);
+}
