+++ IPBot/IPBot/Interfaces/ITweetAnalyserService.cs
@@ -1,9 +1,10 @@
 namespace IPBot.Interfaces;
 
-public interface ITweetService
+public interface ITweetAnalyserService
 {
     Task<List<string>> GetTweetVideoLinksAsync(TweetDetails tweetDetails);
     bool ContentContainsTweetLink(string content);
     TweetDetails GetTweetDetails(string tweetLink);
     string GetFixUpXLink(TweetDetails tweetDetails);
+    Task CheckForTwitterLinksAsync(SocketMessage message);
 }
